+++ Juego/Clases/Piezas/Peon.py
@@ -2,8 +2,8 @@ from .Pieza import Pieza
 
 class Peon(Pieza):
 
-    def __init__(self, color, id, fila, columna):
-        super().__init__('Peon', color, fila, columna)
+    def __init__(self, color, id, fila, columna, movimiento = None):
+        super().__init__('Peon', color, fila, columna, movimiento)
         self.__id__ = id
 
     def __str__(self):
@@ -17,25 +17,30 @@ class Peon(Pieza):
             if self.__color__ == 'blanco':
                 if self.fila == 7:
                     if fila == 6 or fila == 5:
-                        return 'Recto'
+                        self.__movimiento__ = 'Recto' 
+                        return True
                     else:
                         raise ValueError('El peon se puede mover una o dos casillas hacia adelante')
                 elif fila == self.fila - 1:
-                    return 'Recto'
+                    self.__movimiento__ = 'Recto' 
+                    return True
                 else:
                     raise ValueError('El peon se puede mover una casilla hacia adelante')
             elif self.__color__ == 'negro':
                 if self.fila == 2:
                     if fila == 3 or fila == 4:
-                        return 'Recto'
+                        self.__movimiento__ = 'Recto' 
+                        return True
                     else:
                         raise ValueError('El peon se puede mover una o dos casillas hacia adelante')
                 elif fila == self.fila + 1:
-                    return 'Recto'
+                    self.__movimiento__ = 'Recto' 
+                    return True
                 else:
                     raise ValueError('El peon se puede mover una casilla hacia adelante')
         elif abs(self.__fila__ - fila) == 1 and abs(self.__columna__ - columna) == 1:
-            return 'Comer'
+            self.__movimiento__ = 'Comer' 
+            return True
         else:
             raise ValueError('El peon solo se puede mover en línea recta')