+++ Juego/Clases/Piezas/Reina.py
@@ -2,8 +2,8 @@ from .Pieza import Pieza
    
 class Reina(Pieza):
 
-    def __init__(self, color, fila, columna):
-        super().__init__('Reina', color, fila, columna)
+    def __init__(self, color, fila, columna, movimiento = None):
+        super().__init__('Reina', color, fila, columna, movimiento)
 
     def __str__(self):
         if self.__color__ == 'blanco':
@@ -13,8 +13,10 @@ class Reina(Pieza):
 
     def verificar_movimiento(self, fila, columna):
         if self.fila == fila or self.columna == columna:
-            return 'Recto'
+            self.__movimiento__ = 'Recto' 
+            return True
         elif abs(self.__fila__ - fila) == abs(self.__columna__ - columna):
-            return 'Diagonal'
+            self.__movimiento__ = 'Diagonal' 
+            return True
         else:  
             raise ValueError('El movimiento no es valido')