+++ Juego/Clases/Piezas/Rey.py
@@ -2,8 +2,8 @@ from .Pieza import Pieza
 
 class Rey(Pieza):
 
-    def __init__(self, color, fila, columna):
-        super().__init__('Rey', color, fila, columna)
+    def __init__(self, color, fila, columna, movimiento = None):
+        super().__init__('Rey', color, fila, columna, movimiento)
 
     def __str__(self):
         if self.__color__ == 'blanco':
@@ -14,8 +14,10 @@ class Rey(Pieza):
     def verificar_movimiento(self, fila, columna):
         if abs(self.__fila__ - fila) <= 1 and abs(self.__columna__ - columna) <= 1:
             if abs(self.__fila__ - fila) == abs(self.__columna__ - columna):
-                return 'Diagonal'
+                self.__movimiento__ = 'Diagonal' 
+                return True
             else:  
-                return 'Recto'
+                self.__movimiento__ = 'Recto' 
+                return True
         else:  
             raise ValueError('El movimiento no es valido')