+++ Juego/Clases/Piezas/Torre.py
@@ -2,8 +2,8 @@ from .Pieza import Pieza
 
 class Torre(Pieza):
 
-    def __init__(self, color, id, fila, columna):
-        super().__init__('Torre', color, fila, columna)
+    def __init__(self, color, id, fila, columna, movimiento = None):
+        super().__init__('Torre', color, fila, columna, movimiento)
         self.__id__ = id
 
     def __str__(self):
@@ -14,6 +14,7 @@ class Torre(Pieza):
 
     def verificar_movimiento(self, fila, columna):
         if self.__fila__ == fila or self.__columna__ == columna:
-            return 'Recto'
+            self.__movimiento__ = 'Recto'
+            return True
         else:
             raise ValueError('Movimiento no valido')