+++ Juego/Tests/test_reina.py
@@ -14,12 +14,18 @@ class TestReina(unittest.TestCase):
         self.assertEqual(self.reina_negra.__color__, 'negro')
 
     def test_movimiento_recto_valido(self):
-        self.assertEqual(self.reina_blanca.verificar_movimiento(8, 7), 'Recto')  # Misma fila
-        self.assertEqual(self.reina_blanca.verificar_movimiento(5, 4), 'Recto')  # Misma columna
+        movimiento_b = self.reina_blanca.verificar_movimiento(8, 7)
+        self.assertEqual(movimiento_b, True)
+        self.assertEqual(self.reina_blanca.movimiento, 'Recto')
+        
+        movimiento_b2 = self.reina_blanca.verificar_movimiento(8, 7)
+        self.assertEqual(movimiento_b2, True)
+        self.assertEqual(self.reina_blanca.movimiento, 'Recto')
 
     def test_movimiento_diagonal_valido(self):
-        self.assertEqual(self.reina_blanca.verificar_movimiento(7, 3), 'Diagonal')
-        self.assertEqual(self.reina_blanca.verificar_movimiento(6, 6), 'Diagonal')
+        movimiento_b = self.reina_blanca.verificar_movimiento(7, 3)
+        self.assertEqual(movimiento_b, True)
+        self.assertEqual(self.reina_blanca.movimiento, 'Diagonal')
 
     def test_movimiento_invalido(self):
         with self.assertRaises(ValueError):