+++ Juego/Tests/test_rey.py
@@ -4,7 +4,7 @@ from Juego.Clases.Piezas.Rey import Rey
 class TestRey(unittest.TestCase):
 
     def setUp(self):
-        self.rey_negro = Rey('negro', 0, 5)
+        self.rey_negro = Rey('negro', 1, 5)
         self.rey_blanco = Rey('blanco', 8, 5)
         
     def test_inicializacion(self):
@@ -15,9 +15,18 @@ class TestRey(unittest.TestCase):
 
     # Movimientos válidos dentro del rango del rey
     def test_movimiento_valido(self):
-        self.assertEqual(self.rey_blanco.verificar_movimiento(8, 6), 'Recto')  # Movimiento hacia la derecha
-        self.assertEqual(self.rey_blanco.verificar_movimiento(7, 5), 'Recto')  # Movimiento hacia arriba
-        self.assertEqual(self.rey_blanco.verificar_movimiento(7, 6), 'Diagonal')  # Movimiento diagonal
+
+        movimiento_r = self.rey_blanco.verificar_movimiento(8, 6)
+        self.assertEqual(movimiento_r, True)
+        self.assertEqual(self.rey_blanco.movimiento, 'Recto')
+        
+        movimiento_a = self.rey_blanco.verificar_movimiento(7, 5)
+        self.assertEqual(movimiento_a, True)
+        self.assertEqual(self.rey_blanco.movimiento, 'Recto')
+        
+        movimiento_d = self.rey_blanco.verificar_movimiento(7, 6)
+        self.assertEqual(movimiento_d, True)
+        self.assertEqual(self.rey_blanco.movimiento, 'Diagonal')
 
     
     # Movimiento fuera del rango del rey