+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [1.1.0] -01-09-2024
+
+### Modificado:
+- Se modificaron las piezas, agregando el atributo movimiento, el cual indica el movimiento a realizar de la pieza
+- Se empezo a trabajar en el metodo del jaque
+
+
 ## [1.0.6] - 23-08-2024
 
 ### Modificados: